# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  config.Config:
    properties:
      api_keys_enabled:
        description: 'Allow API key authentication via Authorization: Bearer header.'
        type: boolean
      archive:
        properties:
          generate_nfo_files:
            description: Generate Kodi-compatible NFO sidecars for archived videos.
            type: boolean
          generate_sprite_thumbnails:
            description: Generate sprite thumbnails for scrubbing.
            type: boolean
          save_as_hls:
            description: Save as HLS rather than MP4.
            type: boolean
        type: object
      experimental:
        properties:
          better_live_stream_detection_and_cleanup:
            description: '[EXPERIMENTAL] Enable enhanced detection and cleanup.'
            type: boolean
        type: object
      live_check_interval_seconds:
        description: How often in seconds watched channels are checked for live streams.
        type: integer
      livestream:
        properties:
          proxies:
            description: List of proxies for live stream download.
            items:
              $ref: '#/definitions/config.ProxyListItem'
            type: array
          proxy_enabled:
            description: Enable proxy usage.
            type: boolean
          proxy_parameters:
            description: Query parameters for proxy URL.
            type: string
          proxy_whitelist:
            description: Channels exempt from proxy.
            items:
              type: string
            type: array
          reconnect_grace_minutes:
            description: Resume a dropped live stream into the same video if it returns
              within this many minutes (0 = disabled); captures live streams as HLS.
            minimum: 0
            type: integer
          split_duration_minutes:
            description: Split live archives into .ts parts of at most this many minutes
              (0 = no split); captures live streams as HLS.
            minimum: 0
            type: integer
          watch_while_archiving:
            description: Allow watching live streams while archiving them by downloading
              a temporary HLS stream.
            type: boolean
        type: object
      log_retention_days:
        description: Number of days to retain log files.
        type: integer
      notifications:
        allOf:
        - $ref: '#/definitions/config.LegacyNotification'
        description: |-
          Notifications preserves legacy config.json notifications during migration.
          Deprecated: notifications are now stored in the database.
      parameters:
        properties:
          chat_render:
            description: TwitchDownloaderCLI arguments for chat rendering.
            type: string
          twitch_token:
            description: Twitch token for ad-free live streams or subscriber-only
              videos.
            type: string
          video_convert:
            description: FFmpeg arguments for video conversion.
            type: string
          yt_dlp_video:
            description: yt-dlp arguments for video downloads.
            type: string
        type: object
      registration_enabled:
        description: Enable registration.
        type: boolean
      storage_templates:
        allOf:
        - $ref: '#/definitions/config.StorageTemplate'
        description: Storage folder/file templates.
      tasks:
        properties:
          periodic_update_channels:
            description: Enable periodic channel updates - can cause disks to spin
              up as it checks profile images.
            type: boolean
        type: object
      video_check_interval_minutes:
        description: How often in minutes watched channels are checked for new videos.
        type: integer
    type: object
  config.LegacyNotification:
    properties:
      error_enabled:
        type: boolean
      error_template:
        type: string
      error_webhook_url:
        type: string
      is_live_enabled:
        type: boolean
      is_live_template:
        type: string
      is_live_webhook_url:
        type: string
      live_success_enabled:
        type: boolean
      live_success_template:
        type: string
      live_success_webhook_url:
        type: string
      video_success_enabled:
        type: boolean
      video_success_template:
        type: string
      video_success_webhook_url:
        type: string
    type: object
  config.ProxyListItem:
    properties:
      header:
        description: Optional header to send with the proxy request.
        type: string
      proxy_type:
        allOf:
        - $ref: '#/definitions/utils.ProxyType'
        description: Type of proxy to use.
        minLength: 1
      url:
        description: URL of the proxy server.
        minLength: 1
        type: string
    required:
    - proxy_type
    - url
    type: object
  config.StorageTemplate:
    properties:
      channel_folder_template:
        description: 'Template for channel-level directory naming. Available variables:
          {{channel}}, {{channel_id}}, {{channel_display_name}}.'
        type: string
      file_template:
        type: string
      folder_template:
        type: string
    type: object
...
  http.PollTwitchLoginRequest:
    properties:
      device_code:
        maxLength: 255
        type: string
    required:
    - device_code
    type: object
  http.PollTwitchLoginResponse:
    properties:
      status:
        enum:
        - pending
        - authorized
        type: string
      twitch_login:
        description: Account the token belongs to, when Twitch reports it.
        type: string
      twitch_token:
        description: Set once authorized; already saved to the config.
        type: string
    type: object
...
  platform.TwitchDeviceLogin:
    properties:
      device_code:
        type: string
      expires_in:
        description: Seconds before the code expires.
        type: integer
      interval:
        description: Seconds to wait between polls.
        type: integer
      user_code:
        type: string
      verification_uri:
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  utils.ProxyType:
    enum:
    - twitch_hls
    - http
    type: string
    x-enum-varnames:
    - ProxyTypeTwitchHLS
    - ProxyTypeHTTP
...
host: localhost:4000
info:
  contact: {}
  description: |-
    Authentication is handled using JWT tokens. The tokens are set as access-token and refresh-token cookies.
    For information regarding which role is authorized for which endpoint, see the http handler https://github.com/Zibbp/ganymede/blob/main/internal/transport/http/handler.go.
  title: Ganymede API
  version: "1.0"
paths:
...
  /config:
    get:
      consumes:
      - application/json
      description: Get config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.Config'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - ApiKeyCookieAuth: []
      - ApiKeyAuth: []
      summary: Get config
      tags:
      - config
    put:
      consumes:
      - application/json
      description: Update config
      parameters:
      - description: Config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/config.Config'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.Config'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - ApiKeyCookieAuth: []
      - ApiKeyAuth: []
      summary: Update config
      tags:
      - config
  /config/twitch-login:
    post:
      description: Request a code to authorize at the returned Twitch URL; polling
        it then saves the account's Twitch token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/platform.TwitchDeviceLogin'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - ApiKeyCookieAuth: []
      - ApiKeyAuth: []
      summary: Start Twitch login
      tags:
      - config
  /config/twitch-login/poll:
    post:
      consumes:
      - application/json
      description: Check whether the Twitch login code was authorized. Once it is,
        the account's token is saved as the Twitch token.
      parameters:
      - description: Device code from the start of the login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.PollTwitchLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.PollTwitchLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - ApiKeyCookieAuth: []
      - ApiKeyAuth: []
      summary: Poll Twitch login
      tags:
      - config
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'Authorization: Bearer <api_key>. Mint a key via the admin UI.'
    in: header
    name: Authorization
    type: apiKey
  ApiKeyCookieAuth:
    in: cookie
    name: access-token
    type: apiKey
swagger: "2.0"